+++ ports/ios/GameView.h
@@ -5,6 +5,7 @@
 @property(nonatomic) CGRect textRect;
 @property(nonatomic) BOOL textActive;
 @property(nonatomic) BOOL portraitPanels;
+@property(nonatomic) BOOL preserveAspectRatio;
 - (void)updateTextActive:(BOOL)active rect:(CGRect)rect;
 - (void)cancelGameTouch;
 @end